+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.2.0] - 2026-02-10
+* ENH: New user created email template.
+* ENH: Admin created user set password.
+* ENH: Events - Add WYSIWYG event description.
+* ENH: Email Templates - Clickable variables.
+* ENH: Email Templates - Add button to send a test email.
+* BUG: Event Registration - Fix failed to load event false positive.
+* BUG: Event Registration - Styling tweaks.
+
 ## [1.1.0] - 2026-02-08
 * FEA: Add send new email modal for super admins.
 * FEA: Add email resend functionality.